pyproject.toml
[project]
name = "locksmitha"
version = "0.1.0"
description = "Login/authentication service for beanone organization."
authors = [
    { name = "Beanone Team", email = "beanone@example.com" }
]
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "SQLAlchemy>=2.0.0",
    "asyncpg>=0.28.0",
    "keylin",
    "python-dotenv",
    "python-jose[cryptography]",
    "email-validator",
    "fastapi-users[sqlalchemy]>=12.0.0"
]
requires-python = ">=3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/locksmitha"]

[tool.hatch.version]
path = "src/locksmitha/__init__.py"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "UP", "PL", "RUF"]
ignore = ["PLR2004", "B008"]

[tool.ruff.lint.isort]
known-first-party = ["locksmitha"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = """
    -v
    --cov=src/locksmitha
    --cov-report=term-missing
    --cov-report=html
    --cov-config=.coveragerc
"""
pythonpath = ["."]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "unit: marks tests as unit tests (deselect with '-m \"not unit\"')"
]
asyncio_default_fixture_loop_scope = "function"
